improved data functions
splitted the function fetch_data_table into three modules:
1, the functions getting table names are now in 3 modules,
2, tested functions to download batch data for one table
